+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix #2398 - Auras.lua:303: attempt to index a nil value
 * Fix #2380 - PlayerNames.lua:112: table index is secret
 * Fix #2385 - Mounted condition doesn't update if no other icon is tracking player auras.
+* Fix #2386 - attempt to compare local 'start' a secret number value tainted by ForceTaint_Strong
 
 ## v12.0.9
 * Fix #2371 - Corruption treated as Immolate